testing/web-platform/tests/html/semantics/embedded-content/the-iframe-element/move_iframe_in_dom_03.html

<!DOCTYPE html>
<title>moving modified IFRAME in document (original page from server, DOM modification)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/#iframe-load-event-steps">
<iframe src="support/blank.htm"></iframe>
<div id="target"></div>
<script>
onload = function() {
  var ifr = document.getElementsByTagName('iframe')[0];
  ifr.contentDocument.body.appendChild(ifr.contentDocument.createElement('p')).textContent = 'Modified document';
  setTimeout(function() {
    document.getElementById('target').appendChild(ifr);
    setTimeout(function() {
      assert_equals(ifr.contentDocument.body.textContent.indexOf('Modified'), -1);
      done();
    }, 100);
  }, 100);
}
</script>
